Switch to scanline VGA
